added pages for creating and loading charts
